fixed norm but found error for designe input(#73)
