+++ src/init/validate_player.c
@@ -6,7 +6,7 @@
 /*   By: yshimazu <yshimazu@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/11 15:29:07 by yshimazu          #+#    #+#             */
-/*   Updated: 2022/03/11 15:29:07 by yshimazu         ###   ########.fr       */
+/*   Updated: 2022/03/14 11:16:46 by yshimazu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
